Moves condition handling to trait
